move to post-resolution during install
